requirements.txt
flask
pyTelegramBotAPI
pyjwt
pyjwt[crypto]
waitress